+++ src/meltano/api/config.py
@@ -25,6 +25,7 @@ PROJECT_ROOT_DIR = os.path.dirname(API_ROOT_DIR)
 # -----------------
 SQLALCHEMY_ECHO = False
 SQLALCHEMY_TRACK_MODIFICATIONS = False
+SQLALCHEMY_DATABASE_URI = os.getenv("MELTANO_DATABASE_URI")
 
 # Flask-security
 # -----------------
@@ -62,7 +63,6 @@ MAIL_SERVER = "localhost"
 MAIL_PORT = 587
 MAIL_DEFAULT_SENDER = '"Meltano" <bot@metlano.com>'
 
-
 # Flask-Authlib
 # -----------------
 
@@ -78,3 +78,9 @@ RESTFUL_JSON = {}
 # -----------------
 
 EXECUTOR_PROPAGATE_EXCEPTIONS = True
+
+
+# Flask-CORS
+# -----------------
+
+CORS_EXPOSE_HEADERS = ["X-Meltano-Version"]