+++ src/meltano/cli/permissions.py
@@ -26,7 +26,7 @@ def permissions():
 @click.option(
     "--diff", help="Show full diff, both new and existing permissions.", is_flag=True
 )
-@project
+@project()
 def grant(project, db, spec, dry, diff):
     """Grant the permissions provided in the provided specification file."""
     try: