+++ src/webapp/src/flask.js
@@ -9,7 +9,8 @@ module.exports = function() {
     window.FLASK || {
       airflowUrl: process.env.AIRFLOW_URL,
       appUrl: process.env.MELTANO_WEBAPP_URL,
-      isSendAnonymousUsageStats: false
+      isSendAnonymousUsageStats: false,
+      version: 'source'
     }
   )
 }